refactor: extract buildJellyfinUrl into utils/jellyfinUrl.js
